Fixed #28312 -- Made ModelChoiceIterator __len__ more memory-efficient
Instead of loading all QuerySet results in memory, simply COUNT the
number of entries. It adds an extra query when list() or tuple() is
called on the choices (because both call __len__ then __iter__), but
use less memory because the QuerySet results won't be cached. In most
case, the choices will only be iterated onto, meaning that __len__ won't
be called and only a single query will be executed.

Refs #23623